requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
anthropic==0.34.2
httpx==0.25.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
greenlet>=2.0.0
python-dotenv==1.0.0
docker==7.1.0
requests==2.32.3
python-multipart==0.0.6
openai-whisper
